Implement Cart example from ch-02
Removed the old example/demo functions as well as tests. Included Jest
as a dev dependency.

Added the new first-example from Chapter 2: 'What to test and when' and
wrote new tests based on the first-example.

Defined the class Cart which describes a customer's shopping cart, along
with `addToCart` and `removeFromCart` methods.