Refactor routes to use authentication middleware
- Modify server.ts to use authenticationMiddleware function for all
routes starting with "carts"
- Add createUser function to server.test.ts to create valid user before
all "/carts" tests
- Modify "/carts" tests to set authorization headers for every request
- Resolve conflicts between other tests by clearing Users Map before
each test
-Run coverage report

This change implements basic authorization using HTTP headers which
makes the app more robust. Some of the original tests/error handling is
now redundant as the routes should no longer need to obtain 'username'
from params. The authorization headers should be used instead. But this
is a minor flaw and is irrelevant to this project as the project
emphasizes on testing rather than correctness of the backend functions