Connect application to database and refactor tests
- Connect application to database for data storage
- Modify functions to use database instead of global state
- Create Cart and Inventory controllers
- Write test suites for Cart and Inventory controllers
- Refactor all previous tests to use database instead of global state

This is a major change that improves the reliability and scalability of
the application by using a database for data storage, instead of global
state. It also improves the testability of the application by providing
dedicated controllers and test suites for the Cart and Inventory
functionality. All previous tests have been refactored to use the
database, which ensures that the tests are more representative of the
actual behavior of the application.