Add external api route for recipes
- Added new server routes for accessing inventory items, which send
related recipes data back
- Implemented this using theMealDB API
- Described end-to-end tests for verifying the api endpoint

This change adds more features to the final application, and
demonstrates the usage of external API dependencies. It also
demonstrates testing external API endpoints and verifying if the
server is making the right calls to the correct endpoints. Currently the
tests also send api requests to the endpoints which is not ideal as
api requests add costs and lower the incentive to test the application.