Install nock package
- Installed 'nock' module to test API requests
- Renamed the axios mock file to stop it from being used
- Implemented mocking to the inventory external API url using nock
when testing the backend route '/inventory/:item'

This change allows the tests to not be coupled with the library being
used for making requests. The axios mock required strict replication
of the behavior of the axios library. This increases complexity of the
tests and increases maintenance cost. Using nock allows the tests to
determine the servers response rather than its behavior, which increases
reusability of the tests.